fix: useSearch hook 위치 변경
